pyproject.toml
[project]
name = "consumer_bot"
version = "0.1.0"
description = "Default template for PDM package"
authors = [
    {name = "giorno", email = "giorno.xu@baozun.com"},
]
dependencies = [
    "langchain==0.1.13",
    "chromadb>=0.4.24",
    "langchain-openai>=0.1.3",
    "langserve>=0.1.0",
    "FastAPI>=0.110.2",
    "sse-starlette>=2.1.0",
    "pillow>=10.3.0",
]
requires-python = "==3.11.*"
readme = "README.md"
license = {text = "MIT"}

[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"


[tool.pdm]
distribution = true
